Moving onboarding check to separate component (in progress)
