+++ app/src/main/res/values/attrs_onboarding_permissions_checker.xml
@@ -1,0 +1,8 @@
+<resources>
+  <declare-styleable name="OnboardingPermissionsChecker">
+    <attr name="exampleString" format="string" />
+    <attr name="exampleDimension" format="dimension" />
+    <attr name="exampleColor" format="color" />
+    <attr name="exampleDrawable" format="color|reference" />
+  </declare-styleable>
+</resources>
